% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPT-POT-compile.R
\name{compile-cpt}
\alias{compile-cpt}
\alias{as_cptlist}
\alias{compileCPT}
\alias{summary.CPTspec}
\alias{compilePOT}
\alias{print.CPTspec}
\alias{compilePOT}
\alias{compile.POT_rep}
\alias{compile.CPT_rep}
\title{Compile conditional probability tables / cliques potentials.}
\usage{
as_cptlist(x, forceCheck = FALSE)

compilePOT(x)

\method{compile}{POT_rep}(object, ...)

\method{compile}{CPT_rep}(object, ...)
}
\arguments{
\item{x}{To \code{compileCPT} x is a list of conditional
probability tables; to \code{compilePOT}, x is a list of clique
potentials.}

\item{forceCheck}{Controls if consistency checks of the probability
tables should be made.}

\item{object}{A list of potentials or of CPTs.}

\item{...}{Additional arguments; currently not used.}

\item{details}{Controls amount of print out. Mainly for debugging
purposes.}
}
\value{
\code{compileCPT} returns a list of class 'cptspec'
    \code{compilePOT} returns a list of class 'potspec'
}
\description{
Compile conditional probability tables / cliques
    potentials as a preprocessing step for creating a graphical
    independence network
}
\references{
Søren Højsgaard (2012). Graphical Independence Networks
    with the gRain Package for R. Journal of Statistical Software,
    46(10), 1-26.  \url{http://www.jstatsoft.org/v46/i10/}.
}
\seealso{
\code{\link{extractCPT}}, \code{\link{extractPOT}}
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\keyword{utilities}
