% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt-pot-compile.R
\name{components_compile}
\alias{components_compile}
\alias{compileCPT}
\alias{compilePOT}
\title{Compile conditional probability tables / cliques potentials.}
\usage{
compileCPT(x, ..., forceCheck = TRUE)

compilePOT(x, ..., forceCheck = TRUE)
}
\arguments{
\item{x}{To \code{compileCPT} x is a list of conditional
probability tables; to \code{compilePOT}, x is a list of clique
potentials.}

\item{...}{Additional arguments; currently not used.}

\item{forceCheck}{Controls if consistency checks of the probability
tables should be made.}
}
\value{
A list with a class attribute.
}
\description{
Compile conditional probability tables / cliques
potentials as a preprocessing step for creating a graphical
independence network
}
\examples{

## FIXME: compileCPT/compilePOT examples missing.
}
\references{
Søren Højsgaard (2012). Graphical Independence Networks
with the gRain Package for R. Journal of Statistical Software,
46(10), 1-26.  \url{http://www.jstatsoft.org/v46/i10/}.
}
\seealso{
\code{\link{extractCPT}}, \code{\link{extractPOT}}
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\keyword{utilities}
