% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_grain.R
\docType{data}
\name{example_chest}
\alias{example_chest}
\alias{example_chest_cpt}
\title{Chest clinic example}
\description{
Conditional probability tables for the chest clinic example.
}
\examples{

yn   <- c("yes", "no")
a    <- cpt(~asia, values=c(1,99),levels=yn)
t.a  <- cpt(~tub|asia, values=c(5,95,1,99),levels=yn)
s    <- cpt(~smoke, values=c(5,5), levels=yn)
l.s  <- cpt(~lung|smoke, values=c(1,9,1,99), levels=yn)
b.s  <- cpt(~bronc|smoke, values=c(6,4,3,7), levels=yn)
e.lt <- cpt(~either|lung:tub,values=c(1,0,1,0,1,0,0,1),levels=yn)
x.e  <- cpt(~xray|either, values=c(98,2,5,95), levels=yn)
d.be <- cpt(~dysp|bronc:either, values=c(9,1,7,3,8,2,1,9), levels=yn)

chest_cpt <- list(a, t.a, s, l.s, b.s, e.lt, x.e, d.be)
## bn <- grain(compile_cpt(chest_cpt))
}
\keyword{datasets}
