% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_grain.R
\docType{data}
\name{example_grass}
\alias{example_grass}
\alias{example_grass_cpt}
\title{Wet grass example}
\description{
Conditional probability tables for the wet grass example.
}
\examples{
yn <- c("yes", "no")
p.R    <- cpt(~R, values=c(.2, .8), levels=yn)
p.S_R  <- cpt(~S:R, values=c(.01, .99, .4, .6), levels=yn)
p.G_SR <- cpt(~G:S:R, values=c(.99, .01, .8, .2, .9, .1, 0, 1), levels=yn)

grass_cpt <- list(p.R, p.S_R, p.G_SR)
## bn <- grain(compile_cpt(grass_cpt))
}
\keyword{datasets}
