% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{generics}
\alias{generics}
\alias{nodeNames}
\alias{nodeNames.grain}
\alias{nodeStates}
\alias{nodeStates.grain}
\alias{universe}
\alias{universe.grain}
\alias{varNames.grainEvidence_}
\alias{rip.grain}
\alias{vpar.cpt_spec}
\alias{vpar.cpt_grain}
\title{gRain generics}
\usage{
nodeNames(x)

\method{nodeNames}{grain}(x)

nodeStates(x, nodes = nodeNames(x))

\method{nodeStates}{grain}(x, nodes = nodeNames(x))

universe(object, ...)

\method{universe}{grain}(object, ...)

\method{varNames}{grainEvidence_}(x)

\method{rip}{grain}(object, ...)

\method{vpar}{cpt_spec}(object, ...)

\method{vpar}{cpt_grain}(object, ...)
}
\arguments{
\item{x, object}{A relevant object.}

\item{nodes}{Some nodes of the object.}

\item{...}{Additional arguments; currently not used.}
}
\description{
Generic functions etc for the gRain package
}
