\name{gmInstance-class}
\docType{class}
\alias{gmInstance-class}

\title{
  Class "gmInstance" (graphical model instance)
}
\description{
  The \code{gmInstance} class provides a representation of a
  "graphical model instance"
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
%\section{Extends}{
%Class \code{"\linkS4class{oldClass}"}, directly.
%}
\section{Methods}{
  No methods defined with class "gmInstance" in the signature.
}
%\references{ ~put references to the literature/web site here ~ }
\author{Søren Højsgaard, sorenh@agrsci.dk}
%\note{ ~~further notes~~ }

% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%  See Also% \code{\link{makebn}}
%  % 	or \code{\linkS4class{CLASSNAME}} for links to other classes
%}

\examples{
showClass("gmInstance")
}
\keyword{classes}
