% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{grain-generics}
\alias{grain-generics}
\alias{nodeNames}
\alias{nodeNames.grain}
\alias{nodeStates}
\alias{nodeStates.grain}
\alias{pot}
\alias{pot.grain}
\alias{potential}
\alias{potential.grain}
\alias{rip.grain}
\alias{uni}
\alias{uni.grain}
\alias{universe}
\alias{universe.grain}
\alias{varNames.grainEvidence_}
\title{gRain generics}
\usage{
nodeNames(x)

\method{nodeNames}{grain}(x)

nodeStates(x, nodes = nodeNames(x))

\method{nodeStates}{grain}(x, nodes = nodeNames(x))

universe(object, ...)

\method{universe}{grain}(object, ...)

\method{varNames}{grainEvidence_}(x)

\method{rip}{grain}(object, ...)

uni(object)

\method{uni}{grain}(object)

pot(object)

\method{pot}{grain}(object)

potential(object)

\method{potential}{grain}(object)
}
\arguments{
\item{x, object}{A relevant object.}

\item{nodes}{Some nodes of the object.}

\item{...}{Additional arguments; currently not used.}
}
\description{
Generic functions etc for the gRain package
}

