% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRain-package.R
\docType{package}
\name{gRain-package}
\alias{gRain-package}
\alias{gRain}
\title{Overview - the gRain package for graphical independence networks}
\description{
Probability propagation in graphical independence networks, (also known as
probabilistic expert or Bayesian networks).
}
\details{
This package implements graphical independence networks.

The main function for building networks is \code{\link{grain}}.

The function \code{\link{querygrain}} is used for querying independence
networks.

Functions \code{\link{simulate.grain}} and \code{\link{predict.grain}} are
available.

There is a small vignette which illustrates the use of gRain.
}
\references{
Søren Højsgaard (2012). Graphical Independence Networks
with the gRain Package for R. Journal of Statistical Software, 46(10), 1-26.
\url{http://www.jstatsoft.org/v46/i10/}.
}
\author{
Søren Højsgaard \email{sorenh@math.aau.dk}

Maintainer: Søren Højsgaard \email{sorenh@math.aau.dk}
}
\keyword{Internal}
\keyword{These}
\keyword{be}
\keyword{by}
\keyword{called}
\keyword{directly}
\keyword{for}
\keyword{functions}
\keyword{gRain}
\keyword{internal}
\keyword{not}
\keyword{package}
\keyword{should}
\keyword{the}
\keyword{user}
