\name{graphsh-class}
\docType{class}
\alias{graphsh-class}
\alias{ugsh-class}
\alias{dagsh-class}

\title{Class "graphsh"}
\description{The class "graphsh" is a general representation of a graph
  object. "ugsh" and "dagsh" are specializations into for undirected
  graphs and directed acyclic graphs.}
\section{Objects from the Class}{
  Objects can be created by calls of the form \code{new("ugsh", ...)} or
  using the creator functions \code{\link{newugsh}} or \code{\link{newug}}.
  (Similar for "dagsh").
}
%\section{Extends}{
%Class \code{"\linkS4class{oldClass}"}, directly.
%}
%\section{Methods}{
%No methods defined with class "graphsh" in the signature.
%}
%\references{ ~put references to the literature/web site here ~ }
\author{Søren Højsgaard, sorenh@agrsci.dk}
%\note{ ~~further notes~~ }

% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
  \code{\link{newugsh}}, \code{\link{newug}},
  \code{\link{newdagsh}}, \code{\link{newdag}}
  }
\keyword{classes}
