\name{internal-gRain}

\alias{glist2setMAT}  %% FIXME: Should go to gRbase

\alias{update.cpt-grain}
\alias{cpt2} %% FIXME: CRAP NAME
\alias{propagateLS}

\alias{nodeMarginal}
\alias{nodeJoint}
\alias{summary.grain}

\alias{printlist}
\alias{printlist.default}
\alias{printlist.list}
\alias{splitVec}
\alias{splitVec.default}
\alias{splitVec.list}

\alias{print.huginNet}
\alias{print.cptable}
\alias{print.CPTspec}
\alias{print.grain}
\alias{print.grainFinding}
\alias{print.POTspec}

\alias{random_cpt}


% \alias{}
%\alias{}

\title{Internal functions for the gRain package}
\description{These functions should not be called directly by the user}
\keyword{internal}% at least one, from doc/KEYWORDS
