% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptable-object.R
\name{object-cptable}
\alias{object-cptable}
\alias{cpt_domain}
\alias{print.cpt_domain}
\alias{add_cpt}
\alias{add_cpt.cpt_domain}
\alias{add_cpt<-}
\alias{add_cpt<-.cpt_domain}
\alias{mvn_domain}
\alias{print.mvn_domain}
\alias{add_mvn}
\alias{add_mvn.mvn_domain}
\alias{add_mvn<-}
\alias{add_mvn<-.mvn_domain}
\alias{compileMVN}
\title{Object oriented specification of Bayesian network}
\usage{
cpt_domain()

\method{print}{cpt_domain}(x, ...)

add_cpt(object, value)

\method{add_cpt}{cpt_domain}(object, value)

add_cpt(object) <- value

\method{add_cpt}{cpt_domain}(object) <- value

mvn_domain()

\method{print}{mvn_domain}(x, ...)

add_mvn(object, value)

\method{add_mvn}{mvn_domain}(object, value)

add_mvn(object) <- value

\method{add_mvn}{mvn_domain}(object) <- value

compileMVN(bn, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}

\item{object}{An object to be given a value.}

\item{value}{A value to be added to an object.}
}
\description{
FIXME write.
}
\details{
FIXME to be written. This is all experimental and may well change soon.
}
\examples{

yn <- c("yes", "no")
ssp <- list(Rain = yn, Sprinkler = yn, GrassWet = yn)

p.R <- parray("Rain", levels=ssp, values=c(.2, .8))
p.S_R <- parray(c("Sprinkler", "Rain"), levels = ssp,
   values=c(.01, .99, .4, .6))
p.G_SR <- parray(~ GrassWet:Sprinkler:Rain, levels = ssp,
   values=c(.99, .01, .8, .2, .9, .1, 0, 1))

bn <- cpt_domain()
add_cpt(bn) <- p.R
add_cpt(bn) <- p.S_R
add_cpt(bn) <- p.G_SR
grain(compileCPT(bn))

bn <- cpt_domain()
add_cpt(bn) <- cptable(~rain, levels=ssp, values=c(.2, .8))
add_cpt(bn) <- cptable(~sprinkler|rain, levels=ssp, values=c(.01, .99, .4, .6))
add_cpt(bn) <- cptable(~GrassWet|Sprinkler:Rain, levels=ssp, values=c(.99, .01, .8, .2, .9, .1, 0, 1))

}
