% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grain-recent.R
\name{recent}
\alias{recent}
\alias{rip<-}
\alias{rip<-.grain}
\alias{mkcptlist}
\alias{as.cptlist}
\title{Recent additions to gRain}
\usage{
rip(object) <- value
}
\arguments{
\item{object}{An appropriate R object}

\item{value}{A value to be assigned}

\item{...}{Additional arguments}
}
\description{
Recent additions to gRain
}
