% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{simplify_query}
\alias{simplify_query}
\title{Simplify output query to a Bayesian network}
\usage{
simplify_query(b)
}
\arguments{
\item{b}{Result from running querygrain.}
}
\description{
Simplify output query to a Bayesian network to a dataframe provided
that each node has the same levels.
}
