% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{generics}
\alias{generics}
\alias{nodeNames}
\alias{nodeStates}
\alias{nodeNames.grain}
\alias{nodeStates.grain}
\alias{universe}
\alias{universe.grain}
\alias{varNames.grainEvidence_}
\alias{rip.grain}
\alias{uni}
\alias{uni.grain}
\alias{pot}
\alias{pot.grain}
\alias{cpt}
\alias{cpt.cpt_grain}
\alias{cpt<-}
\alias{cpt<-.cpt_grain}
\alias{potential}
\alias{potential.grain}
\alias{vpar.cpt_spec}
\alias{vpar.cpt_grain}
\title{gRain generics}
\usage{
nodeNames(x)

\method{nodeNames}{grain}(x)

nodeStates(x, nodes = nodeNames(x))

\method{nodeStates}{grain}(x, nodes = nodeNames(x))

universe(object, ...)

\method{universe}{grain}(object, ...)

\method{varNames}{grainEvidence_}(x)

\method{rip}{grain}(object, ...)

uni(object)

\method{uni}{grain}(object)

pot(object)

\method{pot}{grain}(object)

cpt(object)

\method{cpt}{cpt_grain}(object)

cpt(object, position) <- value

\method{cpt}{cpt_grain}(object, position) <- value

potential(object)

\method{potential}{grain}(object)

\method{vpar}{cpt_spec}(object, ...)

\method{vpar}{cpt_grain}(object, ...)
}
\arguments{
\item{x, object}{A relevant object.}

\item{nodes}{Some nodes of the object.}

\item{...}{Additional arguments; currently not used.}

\item{position}{Where to insert 'value'}

\item{value}{Value to insert at 'position'}
}
\description{
Generic functions etc for the gRain package
}
