% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{generics}
\alias{generics}
\alias{nodeNames}
\alias{nodeNames.grain}
\alias{nodeStates}
\alias{nodeStates.grain}
\alias{universe}
\alias{universe.grain}
\alias{isCompiled}
\alias{isPropagated}
\alias{vpar.cpt_spec}
\alias{vpar.cpt_grain}
\alias{rip.grain}
\title{gRain generics}
\usage{
nodeNames(object)

\method{nodeNames}{grain}(object)

nodeStates(object, nodes = nodeNames(object))

\method{nodeStates}{grain}(object, nodes = nodeNames(object))

universe(object, ...)

\method{universe}{grain}(object, ...)

isCompiled(object)

isPropagated(object)

\method{vpar}{cpt_spec}(object, ...)

\method{vpar}{cpt_grain}(object, ...)

\method{rip}{grain}(object, ...)
}
\arguments{
\item{object}{A relevant object.}

\item{nodes}{Some nodes of the object.}

\item{...}{Additional arguments; currently not used.}
}
\description{
Generic functions etc for the gRain package
}
