% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{grain-generics}
\alias{grain-generics}
\alias{nodeNames}
\alias{nodeStates}
\alias{nodeNames.grain}
\alias{nodeStates.grain}
\alias{nodeNames.grain}
\alias{nodeStates}
\alias{nodeStates.grain}
\alias{universe}
\alias{universe.grain}
\alias{varNames.grainEvidence_}
\alias{rip.grain}
\alias{uni}
\alias{uni.grain}
\alias{pot}
\alias{pot.grain}
\alias{cpt}
\alias{cpt.grain}
\alias{potential}
\alias{potential.grain}
\alias{vpar.CPTspec}
\alias{vpar.CPTgrain}
\title{gRain generics}
\usage{
nodeNames(x)

\method{nodeNames}{grain}(x)

nodeStates(x, nodes = nodeNames(x))

\method{nodeStates}{grain}(x, nodes = nodeNames(x))

universe(object, ...)

\method{universe}{grain}(object, ...)

\method{varNames}{grainEvidence_}(x)

\method{rip}{grain}(object, ...)

uni(object)

\method{uni}{grain}(object)

pot(object)

\method{pot}{grain}(object)

cpt(object)

\method{cpt}{grain}(object)

potential(object)

\method{potential}{grain}(object)

\method{vpar}{CPTspec}(object, ...)

\method{vpar}{CPTgrain}(object, ...)
}
\arguments{
\item{x, object}{A relevant object.}

\item{nodes}{Some nodes of the object.}

\item{...}{Additional arguments; currently not used.}
}
\description{
Generic functions etc for the gRain package
}
