\name{gRain-package}
\alias{gRain-package}
\alias{gRain}
\docType{package}
\title{
Graphical Independence Networks
}
\description{
A package for propagation in graphical independence
networks.
}
\details{
\tabular{ll}{
Package: \tab gRain\cr
Version: \tab 0.1.0\cr
License: \tab GPL version 2 or newer\cr
Depends: \tab R (>= 2.1.0),methods,gRbase,Rgraphviz,RBGL,doBy\cr
Imports: \tab graph\cr
URL: \tab http://gbi.agrsci.dk/~shd/Public/gRainweb/\cr
Encoding: \tab latin1\cr
Built: \tab R 2.6.1; i386-pc-mingw32; 2008-01-22 00:54:53; windows\cr
}

Index:
\preformatted{
cpt                     Create conditional probability tables (CPTs)
ctab                    Representation of and operations on
                        multidimensional tables
gmInstance-class        Class "gmInstance" (graphical model instance)
graphsh-class           Class "graphsh"
loadHuginNet            Load and save Hugin net files
%mcs                     Operations on undirected and directed acyclic
%                        graphs.
newugsh                 Create undirected and directed graphs
newgmInstance           Graphical Independence Network
predict.compgmInstance
                        Make predictions from a probabilistic network
queryg                  Query a graph
querygm                 Query an independence network
simulate.compgmInstance
                        Simulate from an independence network
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Søren Højsgaard <sorenh@agrsci.dk>

Maintainer: Søren Højsgaard <sorenh@agrsci.dk>
}
\references{
%~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
%~~ directory ~~
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%~~ simple examples of the most important functions ~~
}
