\name{gRain-package}
\alias{gRain-package}
\alias{gRain}
\docType{package}
\title{
Graphical Independence Networks

}
\description{
A package for probability propagation in graphical
independence networks, also known as probabilistic expert
systems (which includes Bayesian networks as a special case).
}
\details{
  \item This package implements graphical independence networks.
  \item The package is still at an experimental stage.
  \item The main function for building networks is
  \code{\link{grain}}
  \item The function \code{\link{querygm}} is used for querying
  independence networks. Functions \code{\link{simulate.grain}} and
  \code{\link{predict.grain}} are available.
%   \item The functions \code{\link{newUG}} and \code{\link{newDAG}}
%   create undirected and directed acyclic graphs. The function
%   \code{\link{queryg}} is used for querying graphs.
  
}
\author{
Søren Højsgaard <sorenh@agrsci.dk>

Maintainer: Søren Højsgaard <sorenh@agrsci.dk>
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~

\keyword{ package }
%\seealso{
%
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
