\name{gRain-package}
\alias{gRain-package}
\alias{gRain}
\docType{package}
\title{
  Overview - the gRain package for graphical independence networks
}
\description{
  Probability propagation in graphical independence networks,
  (also known as probabilistic expert or  Bayesian networks).
}
\details{
  This package implements graphical independence networks.

  The main function for building networks is \code{\link{grain}}.
  
  The function \code{\link{querygrain}} is used for querying
  independence networks.

  Functions \code{\link{simulate.grain}} and
  \code{\link{predict.grain}} are available.

  There is a small vignette which illustrates the use of gRain.
}
\author{
Søren Højsgaard <sorenh at mail dot dk>

Maintainer: Søren Højsgaard <sorenh at mail dot dk>
}

\references{
  Søren Højsgaard (2012). Graphical Independence Networks with the
  gRain Package for R. Journal of Statistical Software, 46(10), 1-26.
  \url{http://www.jstatsoft.org/v46/i10/}.
}


%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~

\keyword{ package }
%\seealso{
%
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
