\name{internal-gRain}

\alias{summary.grain}

\alias{print.huginNet}
\alias{print.cptable}
\alias{print.CPTspec}
\alias{print.grain}
\alias{print.grainFinding}
\alias{print.POTspec}

\alias{update.cpt-grain}
\alias{random_cpt}
\alias{getgrain}

\alias{printlist}
\alias{printlist.default}
\alias{printlist.list}

\alias{splitVec}
\alias{splitVec.default}
\alias{splitVec.list}

%\alias{}
%\alias{}

\title{Internal functions for the gRain package}
\description{These functions should not be called directly by the user}
\keyword{internal}% at least one, from doc/KEYWORDS
