\name{internal-gRain}

\alias{cpt2} %% FIXME: CRAP NAME

\alias{propagateLS}

\alias{as.cumcounts}
\alias{as.cumcounts.data.frame}

\alias{print.grain}
\alias{print.grainFinding}

\alias{checkdag}
\alias{ctabnames}

\alias{nodeMarginal}
\alias{nodeJoint}

\alias{dag2cptspec}
\alias{print.huginNet}

\alias{summary.grain}

\alias{printlist}
\alias{printlist.default}
\alias{printlist.list}
\alias{splitVec}
\alias{splitVec.default}
\alias{splitVec.list}

%\alias{}

\title{Internal functions for the gRain package}
\description{These functions should not be called directly by the user}
\keyword{internal}% at least one, from doc/KEYWORDS
