\name{internal-gRain}



\alias{neigh}
\alias{closure}
\alias{subgraph}
\alias{isComplete}
\alias{isSimplicial}
\alias{simplicialNodes}

\alias{as.data.frame.ctab}
\alias{as.double.ctab}
\alias{as.ctab.numeric}
\alias{print.ugsh}
\alias{print.dagsh}
\alias{print.ptab}

\alias{convertg}
\alias{ancestralGraph}

\alias{mcwhSH}

\alias{as.cumcounts}
\alias{as.cumcounts.data.frame}


\alias{print.gmInstance}
\alias{print.cptspec}
\alias{print.bnevidence}
\alias{print.ripOrder}

\alias{is.subset}
\alias{is.subsetList}
\alias{names2pairs}

\alias{compilegm}
\alias{compilegm.cpt-gmInstance}
\alias{compilegm.dag-gmInstance}
\alias{compilegm.ug-gmInstance}



\alias{propagate}
\alias{propagate.compgmInstance}

\alias{cardOrder}
\alias{checkdag}
\alias{ctabnames}


\alias{maximalSet}
\alias{nodeMarginal}
\alias{nodeJoint}
\alias{propagate}
\alias{vpav}

\alias{dag2cptspec}
\alias{print.huginNet}

\alias{ug2dag}
\alias{summary.gmInstance}
\alias{printlist}
\alias{printlist.default}
\alias{printlist.list}
\alias{splitVec}
\alias{splitVec.default}
\alias{splitVec.list}

\alias{getSlot}
\alias{updateugsh}
%\alias{}

\title{Internal functions for the gRbayesnet package}
\description{These functions should not be called directly by the user}
\keyword{internal}% at least one, from doc/KEYWORDS
