\name{internal-gRain}

%\alias{mendel}

\alias{Curry}

\alias{summary.grain}

\alias{print.huginNet}
\alias{print.cptable}
%%\alias{print.CPTspec}
\alias{print.grain}
\alias{print.grainFinding}
\alias{print.POTspec}

\alias{randomCPT}
\alias{getgrain}
\alias{getgin}

\alias{printlist}
\alias{printlist.default}
\alias{printlist.list}

\alias{splitVec}
\alias{splitVec.default}
\alias{splitVec.list}

\alias{valueLabels.cptable}
\alias{varNames.cptable}
\alias{valueLabels.table}
\alias{varNames.table}


% FIXME: Should go to gRbase
\alias{setSliceValue}
\alias{tableListProd}
\alias{print.grainEvidence_}
\alias{print.grainJEvidence_}

%\alias{setJointEvidence}
%\alias{retractJointEvidence}

\alias{\%>\%}



%%\alias{propagateLS__}
%\alias{}

\title{Internal functions for the gRain package}
\description{These functions should not be called directly by the user}
\keyword{internal}% at least one, from doc/KEYWORDS
