% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grain_internal.R
\name{internal-grain}
\alias{internal-grain}
\alias{getgin}
\alias{getgrain}
\alias{plot.grain}
\alias{printlist}
\alias{splitVec}
\title{Internal functions for the gRain package}
\description{
These functions should not be called directly by the user
}
\keyword{internal}
